requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
Flask
pinecone
python-dotenv
langchain
langchain-openai
langchain-community
beautifulsoup4
pypdf

#-e .
